+++ instrumentation/http/lib/opentelemetry/instrumentation/http/patches/client.rb
@@ -13,6 +13,7 @@ module OpenTelemetry
           def perform(req, options)
             uri = req.uri
             request_method = req.verb.to_s.upcase
+            span_name = create_request_span_name(request_method, uri.path)
 
             attributes = {
               'http.method' => request_method,
@@ -23,7 +24,7 @@ module OpenTelemetry
               'net.peer.port' => uri.port
             }.merge!(OpenTelemetry::Common::HTTP::ClientContext.attributes)
 
-            tracer.in_span("HTTP #{request_method}", attributes: attributes, kind: :client) do |span|
+            tracer.in_span(span_name, attributes: attributes, kind: :client) do |span|
               OpenTelemetry.propagation.inject(req.headers)
               super.tap do |response|
                 annotate_span_with_response!(span, response)
@@ -33,12 +34,27 @@ module OpenTelemetry
 
           private
 
+          def config
+            OpenTelemetry::Instrumentation::HTTP::Instrumentation.instance.config
+          end
+
           def annotate_span_with_response!(span, response)
             return unless response&.status
 
             status_code = response.status.to_i
             span.set_attribute('http.status_code', status_code)
             span.status = OpenTelemetry::Trace::Status.error unless (100..399).include?(status_code.to_i)
+          end
+
+          def create_request_span_name(request_method, request_path)
+            if (implementation = config[:span_name_formatter])
+              updated_span_name = implementation.call(request_method, request_path)
+              updated_span_name.is_a?(String) ? updated_span_name : "HTTP #{request_method}"
+            else
+              "HTTP #{request_method}"
+            end
+          rescue StandardError
+            "HTTP #{request_method}"
           end
 
           def tracer