+++ instrumentation/rack/example/trace_demonstration3.rb
@@ -11,7 +11,7 @@ require 'bundler/setup'
 Bundler.require
 
 ENV['OTEL_TRACES_EXPORTER'] = 'console'
-OpenTelemetry::SDK.configure do |c
+OpenTelemetry::SDK.configure do |c|
   c.use 'OpenTelemetry::Instrumentation::Rack', { response_propagators: [OpenTelemetry::Trace::Propagation::TraceContext::ResponseTextMapPropagator.new] }
 end