+++ instrumentation/mysql2/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History: opentelemetry-instrumentation-mysql2
 
+### v0.23.0 / 2023-04-17
+
+* BREAKING CHANGE: Drop support for EoL Ruby 2.7 
+
+* ADDED: Drop support for EoL Ruby 2.7 
+* FIXED: Ensure encoding errors handled during SQL obfuscation for Trilogy 
+
 ### v0.22.0 / 2023-01-14
 
 * BREAKING CHANGE: Removed deprecated instrumentation options